Reimplement MPRIS genmon
The main improvement over the old implementation is that it now handles
more media players, not only Spotify.

It was problematic when there was multiple media players launched at the
same time, that's why playerctl was replaced with the mpris2controller,
which intelligently controls all launched players.

Minor modifications includes the change in the status format.