Refactor python packages management
`sudo` is system-specific and required only when installing packages
system-wide, which I know, after all that time, that is not the best
practice: system-wide packages should be installed using system package
manager (apt, pacman, brew), not by pip.

Previous version of some of the commands (e.g. installing from the file)
are also from the time when I was newbie Python developer.