Move custom scripts
The primary reason is that ~/.local/bin does not seem to be for
exclusive use: there are other files, installed there probably by OS at
some point.
It makes sometimes difficult to differentiate which files there were put
by me, and which don't.

On the other hand, ~/bin _does_ seem to be for exclusive use: I did not
see the scenario yet, where any package manager or similar software
created anything in this directory.