Add Karabiner configuration
It includes re-mapping of macOS to the keyboard layout in the form of
`Ctrl, Super, Alt` (kind of, because this is implemented as `Cmd, Cmd,
Opt`, with `Ctrl` being in the place of `Caps Lock`).

This is my experiment to provide similar layout on each OS I use (and I
use all the major ones: Windows, Linux, macOS). Previously I tried to
mimick the layout of macOS (`Ctrl, Opt, Cmd`, with `Cmd` being used for
things like copy & paste) on Linux, which was PITA, especially since I
didn't do that on Windows (to prevent messing up controls in video
games, which is currently the _only_ thing I'm using Windows for), and I
regularly mistyped copy & paste on rare occasions when I had to use
those on Windows or use Linux tty (where I didn't remapped the keys,
too, because it's not as simple as `Xmodmap` hacks).

The new approach seem to be much more satisfying. Other Mac users never
use my MacBook, and vice versa, so there's no issue with that.