Update Keychron K14 configuration
Adjust the keyboard being switched to Win/Android mode on macOS.

This is not what is recommended by the Keychron itself, but I wanted to
experiment if it will make regular switching between MacBook and
Windows/Linux-driven laptop, and Android-based phone more convenient.
The Keychron itself does not recommend constantly switching the mode, so
there's a good official reason for that, at least.