Add commands for running unkillable background jobs
"unkillable" as in "they aren't killed when exiting terminal".

This includes the old good `nohup` command, which I've used in the past,
but (surprisingly) never documented, and GNU Screen, which I'm not
typically using (on my machine I prefer `tmux`), but not every machine
has `tmux` installed, and as such, it would be especially useful to have
documented that basic functionality if I ever need it.