Remove outdated/redundant packages
Some of those were removed experimentally (I'm not 100% sure I don't
need them, I just know I don't explicitly use them). If anything breaks,
I'll rollback this change.

My theory is that when transitioning from old MacBook to the new one, I
probably dumped a list of installed packages on the former and installed
them on the latter, and for some reason the dump contained implicit
dependencies which I installed explicitly. That's just a guess, tho.
I've removed some of those I'm well-aware I don't use explicitly, and
we'll see if something breaks or not.

Both `imagemagick` and `libheif` are special case because they are
dependencies of `zbar`, which I'm using, but for some reason they were
marked as explicitly installed. I guess the reason might be similar to
the above one.
I uninstalled them _and_ `zbar` and then reinstalled only `zbar` (which
installed the other two as dependencies, not explicit packages).