Add `vim-LanguageTool` plugin
LanguageTool is a spelling and grammar checker similar to Grammarly, but
it additionally supports many more languages than English (including
_Polish_ and regional variants of English) and offers local offline
server that can be integrated into Vim using the above-mentioned plugin.
I'm adding the plugin **experimentally** to see if it will serve me well
when writing short stories in Polish, as opposed to manually
copy-pasting snippet by snippet (due to the character limit) to the
web-based LanguageTool, and as a complementary tool or potential
alternative to already used Grammarly (which I'm using via the browser
extension).