Install package for retrieving currently playing song
And add the useful alias for it.

This is because macOS does not support Linux's MPRIS interface, so I
cannot use `mpris2controller` (which I use on my Linux machine).

In theory, macOS allows to use `Fn + c` to pop-up the Control Center
which shows currently playing songs, however, the shortcut does not work
when my Keychron K14 keyboard is switched to Win/Android mode[^1].
This CLI tool is a nice alternative.

[^1]: See 88f62a881df6e2e9e59b2e139a1dc6642863bd99.