Remove global blame ignore file
Introduced in b9fa537ab1c98544b7982848189bf1abb30d1500.

I found out this configuration is broken when there is _no_ such file in
a project, so this basically forces me to put `.git-blame-ignore-revs`
in every project where I'm using `git`.
This is not practical, so I'm removing it from the global configuration,
and instead I'll define it in a project-specific (directory-specific)
git configuration for projects where I use it.