Fix simple/complex status
There was a weird issue where if I was on a non-macOS machine, where the
simple status was supposed to be enabled by default, it wasn't until I
reloaded the tmux configuration.

A long research in the tmux man pages didn't really answer why; instead
I experimented with many different approaches to handling the
environment variable and then the if-else condition (`%if`, `if-shell`,
using quotes, using braces, setting the environment variable and
unsetting it vs. setting it with true or false, etc etc).
The current implementation is what I found that works best.

Additionally, I added a check if the `TMUX_SIMPLE_STATUS` is already
set, otherwise it wouldn't be respected if I want to force the simple
status on macOS or the complex status on non-macOS: the current solution
simply overwrite it on reloading the configuration.