Improve mappings for deleting trailing whitespaces
Previously, I only had mapping to do this in a whole file, which wasn't
compatible with markdown files.

Now I can easily do this for a single line or for a visual selection.
To do this for a whole file, I can just highlight the whole file with
visual selection using `ggVG`.