Sort the Homebrew packages
Separately by tap, brew and cask, to preserve the order of these "types".

This was motivated by the fact, that the order of packages is constantly
changing when dumping the `.Brewfile`, in an arbitrary way, which leads
to confusing and impractical diffs.

The disadvantage of the new approach is that I had to lose the
descriptions, but maybe it will be a worthwhile compromise.