Install missing `black`
It was present in the homebrew's directory on `phthia` despite not being
installed.
I have no idea how that worked.

Maybe it was some kind of anomaly after migrating from the previous
laptop to `phthia`?
I don't know.

After moving to `muspelheim` (the new laptop) it didn't work anymore, so
I had to install it.