Add /etc/shells (kind of)
Obviously, it's not possible/practical to commit a file from root-owned
`/etc` directory, so there's a semi-sophisticated mechanism for
installing a home-owned copy to the `/etc`.

This is also the 1st attempt at generalizing the logic for such files.
Previously, I did a lightdm-specific script (see `~/lightdm`), which is
not flexible if there are more `/etc` files than that one (`/etc/shells`
being the 1st such file in the history of my dotfiles, so this is the
1st time I'm considering the alternative solution).