Migrate to macOS-vendored `jq`
I found out today that macOS vendors `jq` now:

```
$ command -v jq
/opt/homebrew/bin/jq
$ jq --version
jq-1.7.1
$ brew uninstall jq
Uninstalling /opt/homebrew/Cellar/jq/1.7.1... (20 files, 1.4MB)
$ command -v jq
/usr/bin/jq
$ jq --version
jq-1.7.1-apple
```

So I'm experimentally uninstalling `jq` from homebrew.

The advantage of that is that I can remove dependency of homebrew from
my local crontab (which is a part of this commit), and therefore,
generalize it to be system-agnostic (previously it was macOS-specific).