Unify indents
So far it was a mix of 2- and 4-characters long indents.
The goal of this commit is to unify them.

I decided to go with 2-characters long indents because the result looks
subjectively better; if I use 4-characters long indents, then a
significant number of the code gets longer than 80 columns.