Update gtk3 bookmarks
I don't use a bookmark to `comics` anymore, and I didn't need a bookmark
to `education` for a long long time at this point.

I removed the `recent` bookmark, as Thunar provides this by default now,
with no way to disable it, so there were duplicated instances of this
bookmark.